Separate out syscall definitions
POSIX signatures for syscalls are often different than the actual
kernel implementation. Sometimes the signedness/constness of an argument
changes, sometimes the signature is completely different, and sometimes
the exact name of the syscall is different. Account for this difference
by specifiying a different set of definitions for the linux kernel
implementation, and add shims where necessary (e.g preadv/pwritev/prlimit).