Use the "no spec" helper with all loads using untrusted addresses
Every load using an address derived from an untrusted source must be
filtered through a nospec helper. We can't use normal memcpy() because
the full extent of the load may extend into secure memory even if the
base address is filtered through the nospec helper. So we use a
custom memcpy that passes the address of each load through to the
nospec helper to render malicious loads benign on mis-speculations.